Split hashmap into .c and .h file
