+++ src/sundials/sundials_hashmap.c
@@ -17,15 +17,13 @@
  * the hash map upon its destruction.
  * -----------------------------------------------------------------*/
 
-#ifndef _SUNDIALS_HASHMAP_H
-#define _SUNDIALS_HASHMAP_H
-
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
 #include "sundials/sundials_errors.h"
 #include "sundials/sundials_types.h"
+#include "sundials_hashmap_impl.h"
 
 static const uint64_t HASH_PRIME        = 14695981039346656037U;
 static const uint64_t HASH_OFFSET_BASIS = 1099511628211U;
@@ -45,23 +43,6 @@ static uint64_t fnv1a_hash(const char* str)
   return hash;
 }
 
-typedef struct _SUNHashMapKeyValue* SUNHashMapKeyValue;
-
-struct _SUNHashMapKeyValue
-{
-  const char* key;
-  void* value;
-};
-
-typedef struct _SUNHashMap* SUNHashMap;
-
-struct _SUNHashMap
-{
-  int size;     /* current number of entries */
-  int max_size; /* max number of entries */
-  SUNHashMapKeyValue* buckets;
-};
-
 /*
   This function creates a new SUNHashMap object allocated to hold
   up to 'max_size' entries.
@@ -74,7 +55,7 @@ struct _SUNHashMap
   **Returns:**
     * A SUNErrCode indicating success or a failure
  */
-static SUNErrCode SUNHashMap_New(int max_size, SUNHashMap* map)
+SUNErrCode SUNHashMap_New(int max_size, SUNHashMap* map)
 {
   int i;
 
@@ -115,8 +96,7 @@ static SUNErrCode SUNHashMap_New(int max_size, SUNHashMap* map)
   **Returns:**
     * A SUNErrCode indicating success or a failure
  */
-static SUNErrCode SUNHashMap_Destroy(SUNHashMap* map,
-                                     void (*freevalue)(void* ptr))
+SUNErrCode SUNHashMap_Destroy(SUNHashMap* map, void (*freevalue)(void* ptr))
 {
   int i;
 
@@ -159,9 +139,8 @@ static SUNErrCode SUNHashMap_Destroy(SUNHashMap* map,
     * ``>=0`` -- the index at which the iteration stopped
     * ``<-1`` -- an error occurred
  */
-static int SUNHashMap_Iterate(SUNHashMap map, int start,
-                              int (*yieldfn)(int, SUNHashMapKeyValue, void*),
-                              void* ctx)
+int SUNHashMap_Iterate(SUNHashMap map, int start,
+                       int (*yieldfn)(int, SUNHashMapKeyValue, void*), void* ctx)
 {
   int i;
 
@@ -201,7 +180,7 @@ static int sunHashMapLinearProbeInsert(int idx, SUNHashMapKeyValue kv, void* ctx
     * ``-1`` -- an error occurred
     * ``-2`` -- the map is full
  */
-static int SUNHashMap_Insert(SUNHashMap map, const char* key, void* value)
+int SUNHashMap_Insert(SUNHashMap map, const char* key, void* value)
 {
   int idx;
   int retval;
@@ -264,7 +243,7 @@ static int sunHashMapLinearProbeGet(int idx, SUNHashMapKeyValue kv, void* key)
     * ``-1`` -- an error occurred
     * ``-2`` -- key not found
  */
-static int SUNHashMap_GetValue(SUNHashMap map, const char* key, void** value)
+int SUNHashMap_GetValue(SUNHashMap map, const char* key, void** value)
 {
   int idx;
   int retval;
@@ -308,8 +287,8 @@ static int SUNHashMap_GetValue(SUNHashMap map, const char* key, void** value)
   **Returns:**
     * A SUNErrCode indicating success or a failure
  */
-static SUNErrCode SUNHashMap_Sort(SUNHashMap map, SUNHashMapKeyValue** sorted,
-                                  int (*compar)(const void*, const void*))
+SUNErrCode SUNHashMap_Sort(SUNHashMap map, SUNHashMapKeyValue** sorted,
+                           int (*compar)(const void*, const void*))
 {
   int i;
 
@@ -339,8 +318,7 @@ static SUNErrCode SUNHashMap_Sort(SUNHashMap map, SUNHashMapKeyValue** sorted,
     * A SUNErrCode indicating success or a failure
  */
 #if SUNDIALS_MPI_ENABLED
-static SUNErrCode SUNHashMap_Values(SUNHashMap map, void*** values,
-                                    size_t value_size)
+SUNErrCode SUNHashMap_Values(SUNHashMap map, void*** values, size_t value_size)
 {
   int i;
   int count = 0;
@@ -358,6 +336,4 @@ static SUNErrCode SUNHashMap_Values(SUNHashMap map, void*** values,
 
   return SUN_SUCCESS;
 }
-#endif
-
 #endif
