+++ src/sundials/sundials_logger_impl.h
@@ -21,7 +21,7 @@
 #include <sundials/sundials_logger.h>
 #include <sundials/sundials_types.h>
 
-#include "sundials_hashmap.h"
+#include "sundials_hashmap_impl.h"
 #include "sundials_utils.h"
 
 #define SUNDIALS_LOGGING_ERROR   1
